e2e_test_framework/requirements.txt
playwright>=1.44.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
